+++ src/ast/BlockNode.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 public record BlockNode(List<Node> statements) implements Node {
     @Override
-    public void execute(Map<String, Integer> localVars, Map<String, Integer> globalVars) throws EvalError {
+    public void execute(Map<String, Long> localVars, Map<String, Long> globalVars) throws EvalError {
         for (Node statement : statements) {
             statement.execute(localVars, globalVars);
         }
