+++ src/ast/DoneNode.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 public class DoneNode implements Node {
     @Override
-    public void execute(Map<String, Integer> localVars, Map<String, Integer> globalVars) throws EvalError {
+    public void execute(Map<String, Long> localVars, Map<String, Long> globalVars) throws EvalError {
         // โยน Exception พิเศษเพื่อให้ระบบประเมินผลหยุดทำงานในเทิร์นนี้
         throw new DoneException();
     }
