+++ src/ast/IntLit.java
@@ -2,9 +2,9 @@ package ast;
 
 import java.util.Map;
 
-public record IntLit(int val) implements Expr {
+public record IntLit(long val) implements Expr {
     @Override
-    public int eval(Map<String, Integer> localVars, Map<String, Integer> globalVars) {
+    public long eval(Map<String, Long> localVars, Map<String, Long> globalVars) {
         return val; // ส่งค่าตัวเลขกลับไปตรงๆ ไม่ต้องคำนวณหรือค้นหาใน Map
     }
 }
