+++ src/ast/MoveNode.java
@@ -10,12 +10,12 @@ public record MoveNode(String directionStr) implements Node {
     private static final DirectionAdapter adapter = new StringToDirectionAdapter();
 
     @Override
-    public void execute(Map<String, Integer> localVars, Map<String, Integer> globalVars) throws EvalError {
+    public void execute(Map<String, Long> localVars, Map<String, Long> globalVars) throws EvalError {
         // ขั้นตอน 1: Adapt ข้อมูล
         DIRECTION dir = adapter.adapt(directionStr);
 
         // ขั้นตอน 2: Business Logic
-        GameState.pay(1);
+        GameState.pay(1L);
         GameState.move(dir.name());
     }
 }
