+++ src/ast/WhileNode.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 public record WhileNode(Expr condition, Node statement) implements Node {
     @Override
-    public void execute(Map<String, Integer> localVars, Map<String, Integer> globalVars) throws EvalError {
+    public void execute(Map<String, Long> localVars, Map<String, Long> globalVars) throws EvalError {
         int counter = 0;
         // ทำงานเมื่อ condition > 0 และวนไม่เกิน 10,000 รอบ
         while (condition.eval(localVars, globalVars) > 0 && counter < 10000) {
